backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
slack_sdk==3.33.4
aiohttp==3.10.10
pydantic==2.9.2

# LLM (optional)
openai==1.57.2

# AWS (optional for real auto-fix)
boto3==1.35.60
botocore==1.35.60

# Image/OCR (optional; code runs without system tesseract present)
pillow==10.4.0
pytesseract==0.3.13
opencv-python-headless==4.10.0.84
